Create a "sidebar stack" to keep track of what content is displayed
This simplifies some of the tray/selected course logic, as well as
moving the concept of the "current course" from a global state
variable to context calculated from the stack directly.  This will
allow us to have a history of selected courses as needed, plus puts
things in the more realistic 'app' namespace.